Revert "Created a discovery handler using asyncio (#46)"
This reverts commit fe0fa1561b4e57336c5ab57acb29a955909bf959.
