+++ .bumpversion.cfg
@@ -5,8 +5,6 @@ current_version = 0.1.34
 
 [bumpversion:file:msmart/client.py]
 
-[bumpversion:file:msmart/const.py]
-
 [bumpversion:file:msmart/lan.py]
 
 [bumpversion:file:msmart/command.py]
@@ -16,8 +14,6 @@ current_version = 0.1.34
 [bumpversion:file:msmart/device.py]
 
 [bumpversion:file:msmart/packet_builder.py]
-
-[bumpversion:file:msmart/scanner.py]
 
 [bumpversion:file:msmart/security.py]
