+++ msmart/lan.py
@@ -2,8 +2,7 @@
 import logging
 import socket
 import time
-from msmart.const import MSGTYPE_ENCRYPTED_REQUEST, MSGTYPE_HANDSHAKE_REQUEST
-from msmart.security import security
+from msmart.security import security, MSGTYPE_HANDSHAKE_REQUEST, MSGTYPE_ENCRYPTED_REQUEST
 
 VERSION = '0.1.34'
