+++ msmart/packet_builder.py
@@ -1,7 +1,8 @@
-import datetime
+
 import logging
 from msmart.command import base_command
 from msmart.security import security
+import datetime
 
 VERSION = '0.1.34'
